Finished.
The two spreadsheets contain the exact same data. pricedata2.xlsx is
simply ordered in a different way so that bokeh would have an easier
time processing it into a dataframe.

Development was done in the Jupyter notebook, but the .py file was
required to build the website due to a current bug with Bokeh's
output_file function and Jupyter.